+++ rename_chain/main.go
@@ -132,6 +132,8 @@ func main() {
 		os.Exit(1)
 	}
 
+	_ = os.RemoveAll("rename_chain")
+
 	ogGitHubWithoutScheme := strings.TrimSuffix(strings.Split(EvermintOg_GitHubRepo, "://")[1], ".git")
 	ogGoModule := fmt.Sprintf("%s/v12", ogGitHubWithoutScheme)
 	splOgGitHub := strings.Split(ogGitHubWithoutScheme, "/")
@@ -187,7 +189,11 @@ func main() {
 
 	patternMarker := regexp.MustCompile(`marker\.(\w+)\("(\w+)"\)`)
 	for _, goFile := range goFiles {
-		if !strings.HasSuffix(goFile, "_test.go") {
+		if strings.HasSuffix(goFile, "_test.go") {
+			// ok
+		} else if strings.HasSuffix(goFile, "integration_test_util/accounts.go") {
+			// ok
+		} else {
 			continue
 		}
 
@@ -216,6 +222,8 @@ func main() {
 			hrp := constants.Bech32Prefix
 			if strings.HasPrefix(ogBech32Address, EvermintOg_Bech32PrefixValAddr) {
 				hrp = constants.Bech32PrefixValAddr
+			} else if strings.HasPrefix(ogBech32Address, EvermintOg_Bech32PrefixValConsAddr) {
+				hrp = constants.Bech32PrefixConsAddr
 			}
 
 			if match[1] == "ReplaceAbleAddress" {
@@ -423,8 +431,9 @@ const (
 	EvermintOg_DisplayDenom = "ether"
 	EvermintOg_SymbolDenom  = "ETH"
 
-	EvermintOg_Bech32Prefix        = "evm"
-	EvermintOg_Bech32PrefixValAddr = "evmvaloper"
+	EvermintOg_Bech32Prefix            = "evm"
+	EvermintOg_Bech32PrefixValAddr     = "evmvaloper"
+	EvermintOg_Bech32PrefixValConsAddr = "evmvalcons"
 
 	EvermintOg_MainnetChainID = "evermint_90909"
 	EvermintOg_TestnetChainID = "evermint_80808"