+++ rpc/backend/blocks_test.go
@@ -2,10 +2,15 @@ package backend
 
 import (
 	"fmt"
+	"github.com/cometbft/cometbft/libs/log"
 	"math/big"
 
 	"cosmossdk.io/math"
 
+	"github.com/EscanBE/evermint/v12/rpc/backend/mocks"
+	ethrpc "github.com/EscanBE/evermint/v12/rpc/types"
+	utiltx "github.com/EscanBE/evermint/v12/testutil/tx"
+	evmtypes "github.com/EscanBE/evermint/v12/x/evm/types"
 	"github.com/cometbft/cometbft/abci/types"
 	tmrpctypes "github.com/cometbft/cometbft/rpc/core/types"
 	tmtypes "github.com/cometbft/cometbft/types"
@@ -14,12 +19,6 @@ import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/trie"
-	"google.golang.org/grpc/metadata"
-
-	"github.com/EscanBE/evermint/v12/rpc/backend/mocks"
-	ethrpc "github.com/EscanBE/evermint/v12/rpc/types"
-	utiltx "github.com/EscanBE/evermint/v12/testutil/tx"
-	evmtypes "github.com/EscanBE/evermint/v12/x/evm/types"
 )
 
 func (suite *BackendTestSuite) TestBlockNumber() {
@@ -30,41 +29,34 @@ func (suite *BackendTestSuite) TestBlockNumber() {
 		expPass        bool
 	}{
 		{
-			"fail - invalid block header height",
-			func() {
-				var header metadata.MD
-				height := int64(1)
-				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterParamsInvalidHeight(queryClient, &header, height)
+			name: "pass - indexer indexed up to block 1",
+			registerMock: func() {
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, 1)
 			},
-			0x0,
-			false,
+			expBlockNumber: 0x1,
+			expPass:        true,
 		},
 		{
-			"fail - invalid block header",
-			func() {
-				var header metadata.MD
-				height := int64(1)
-				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterParamsInvalidHeader(queryClient, &header, height)
+			name: "pass - indexer indexed up to block 3",
+			registerMock: func() {
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, 3)
 			},
-			0x0,
-			false,
+			expBlockNumber: 0x3,
+			expPass:        true,
 		},
 		{
-			"pass - app state header height 1",
-			func() {
-				var header metadata.MD
-				height := int64(1)
-				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterParams(queryClient, &header, height)
+			name: "fail - indexer returns error",
+			registerMock: func() {
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetLastRequestIndexedBlockErr(indexer)
 			},
-			0x1,
-			true,
+			expPass: false,
 		},
 	}
 	for _, tc := range testCases {
-		suite.Run(fmt.Sprintf("Case %s", tc.name), func() {
+		suite.Run(tc.name, func() {
 			suite.SetupTest() // reset test and queries
 			tc.registerMock()
 
@@ -85,7 +77,8 @@ func (suite *BackendTestSuite) TestGetBlockByNumber() {
 		blockRes *tmrpctypes.ResultBlockResults
 		resBlock *tmrpctypes.ResultBlock
 	)
-	msgEthereumTx, bz := suite.buildEthereumTx()
+	msgEthereumTx, _ := suite.buildEthereumTx()
+	msgEthereumTx, bz := suite.signMsgEthTx(msgEthereumTx)
 
 	testCases := []struct {
 		name         string
@@ -166,6 +159,11 @@ func (suite *BackendTestSuite) TestGetBlockByNumber() {
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFee(queryClient, baseFee)
 				RegisterValidatorAccount(queryClient, validator)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
 			},
 			false,
 			true,
@@ -188,9 +186,69 @@ func (suite *BackendTestSuite) TestGetBlockByNumber() {
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFee(queryClient, baseFee)
 				RegisterValidatorAccount(queryClient, validator)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetByTxHash(indexer, msgEthereumTx.AsTransaction().Hash(), height)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
 			},
 			false,
 			true,
+		},
+		{
+			name:        "fail - indexer returns error when fetching tx",
+			blockNumber: ethrpc.BlockNumber(1),
+			fullTx:      false,
+			baseFee:     sdk.NewInt(1).BigInt(),
+			validator:   sdk.AccAddress(utiltx.GenerateAddress().Bytes()),
+			tx:          msgEthereumTx,
+			txBz:        bz,
+			registerMock: func(blockNum ethrpc.BlockNumber, baseFee math.Int, validator sdk.AccAddress, txBz []byte) {
+				height := blockNum.Int64()
+				client := suite.backend.clientCtx.Client.(*mocks.Client)
+				resBlock, _ = RegisterBlock(client, height, txBz)
+				blockRes, _ = RegisterBlockResults(client, blockNum.Int64())
+				RegisterConsensusParams(client, height)
+
+				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
+				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetByTxHashErr(indexer, msgEthereumTx.AsTransaction().Hash())
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
+			},
+			expNoop: false,
+			expPass: false,
+		},
+		{
+			name:        "pass - indexer returns error when get latest block number",
+			blockNumber: ethrpc.BlockNumber(1),
+			fullTx:      false,
+			baseFee:     sdk.NewInt(1).BigInt(),
+			validator:   sdk.AccAddress(utiltx.GenerateAddress().Bytes()),
+			tx:          msgEthereumTx,
+			txBz:        bz,
+			registerMock: func(blockNum ethrpc.BlockNumber, baseFee math.Int, validator sdk.AccAddress, txBz []byte) {
+				height := blockNum.Int64()
+				client := suite.backend.clientCtx.Client.(*mocks.Client)
+				resBlock, _ = RegisterBlock(client, height, txBz)
+				blockRes, _ = RegisterBlockResults(client, blockNum.Int64())
+				RegisterConsensusParams(client, height)
+
+				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
+				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetByTxHash(indexer, msgEthereumTx.AsTransaction().Hash(), height)
+				RegisterIndexerGetLastRequestIndexedBlockErr(indexer)
+			},
+			expNoop: false,
+			expPass: true,
 		},
 	}
 	for _, tc := range testCases {
@@ -228,7 +286,8 @@ func (suite *BackendTestSuite) TestGetBlockByHash() {
 		blockRes *tmrpctypes.ResultBlockResults
 		resBlock *tmrpctypes.ResultBlock
 	)
-	msgEthereumTx, bz := suite.buildEthereumTx()
+	msgEthereumTx, _ := suite.buildEthereumTx()
+	msgEthereumTx, bz := suite.signMsgEthTx(msgEthereumTx)
 
 	block := tmtypes.MakeBlock(1, []tmtypes.Tx{bz}, nil, nil)
 
@@ -311,6 +370,11 @@ func (suite *BackendTestSuite) TestGetBlockByHash() {
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFee(queryClient, baseFee)
 				RegisterValidatorAccount(queryClient, validator)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
 			},
 			false,
 			true,
@@ -334,9 +398,38 @@ func (suite *BackendTestSuite) TestGetBlockByHash() {
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFee(queryClient, baseFee)
 				RegisterValidatorAccount(queryClient, validator)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetByTxHash(indexer, msgEthereumTx.AsTransaction().Hash(), height)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
 			},
 			false,
 			true,
+		},
+		{
+			name:      "success - indexer returns error",
+			hash:      common.BytesToHash(block.Hash()),
+			fullTx:    true,
+			baseFee:   sdk.NewInt(1).BigInt(),
+			validator: sdk.AccAddress(utiltx.GenerateAddress().Bytes()),
+			registerMock: func(hash common.Hash, baseFee math.Int, validator sdk.AccAddress, txBz []byte) {
+				height := int64(1)
+				client := suite.backend.clientCtx.Client.(*mocks.Client)
+				resBlock, _ = RegisterBlockByHash(client, hash, txBz)
+
+				blockRes, _ = RegisterBlockResults(client, height)
+				RegisterConsensusParams(client, height)
+
+				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
+				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetLastRequestIndexedBlockErr(indexer)
+			},
+			expPass: true,
 		},
 	}
 	for _, tc := range testCases {
@@ -562,25 +655,33 @@ func (suite *BackendTestSuite) TestTendermintBlockByNumber() {
 			true,
 		},
 		{
-			"fail - blockNum < 0 with app state height error",
+			"fail - blockNum < 0 with indexer returns error",
 			ethrpc.BlockNumber(-1),
 			func(_ ethrpc.BlockNumber) {
-				var header metadata.MD
-				appHeight := int64(1)
-				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterParamsError(queryClient, &header, appHeight)
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetLastRequestIndexedBlockErr(indexer)
 			},
 			false,
 			false,
 		},
 		{
-			"pass - blockNum < 0 with app state height >= 1",
+			name:        "fail - blockNum < 0 with indexer returns error",
+			blockNumber: ethrpc.BlockNumber(-1),
+			registerMock: func(_ ethrpc.BlockNumber) {
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetLastRequestIndexedBlockErr(indexer)
+			},
+			found:   false,
+			expPass: false,
+		},
+		{
+			"pass - blockNum < 0 with indexed height >= 1",
 			ethrpc.BlockNumber(-1),
 			func(blockNum ethrpc.BlockNumber) {
-				var header metadata.MD
 				appHeight := int64(1)
-				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterParams(queryClient, &header, appHeight)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, appHeight)
 
 				tmHeight := appHeight
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
@@ -887,7 +988,8 @@ func (suite *BackendTestSuite) TestBlockBloom() {
 }
 
 func (suite *BackendTestSuite) TestGetEthBlockFromTendermint() {
-	msgEthereumTx, bz := suite.buildEthereumTx()
+	msgEthereumTx, _ := suite.buildEthereumTx()
+	msgEthereumTx, bz := suite.signMsgEthTx(msgEthereumTx)
 	emptyBlock := tmtypes.MakeBlock(1, []tmtypes.Tx{}, nil, nil)
 
 	testCases := []struct {
@@ -920,6 +1022,11 @@ func (suite *BackendTestSuite) TestGetEthBlockFromTendermint() {
 
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				RegisterConsensusParams(client, height)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
 			},
 			false,
 			true,
@@ -944,6 +1051,12 @@ func (suite *BackendTestSuite) TestGetEthBlockFromTendermint() {
 
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				RegisterConsensusParams(client, height)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetByTxHash(indexer, msgEthereumTx.AsTransaction().Hash(), height)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
 			},
 			true,
 			true,
@@ -968,6 +1081,12 @@ func (suite *BackendTestSuite) TestGetEthBlockFromTendermint() {
 
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				RegisterConsensusParams(client, height)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetByTxHash(indexer, msgEthereumTx.AsTransaction().Hash(), height)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
 			},
 			true,
 			true,
@@ -992,6 +1111,12 @@ func (suite *BackendTestSuite) TestGetEthBlockFromTendermint() {
 
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				RegisterConsensusParamsError(client, height)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetByTxHash(indexer, msgEthereumTx.AsTransaction().Hash(), height)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
 			},
 			true,
 			true,
@@ -1022,6 +1147,11 @@ func (suite *BackendTestSuite) TestGetEthBlockFromTendermint() {
 
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				RegisterConsensusParams(client, height)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
 			},
 			false,
 			true,
@@ -1046,6 +1176,12 @@ func (suite *BackendTestSuite) TestGetEthBlockFromTendermint() {
 
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				RegisterConsensusParams(client, height)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetByTxHash(indexer, msgEthereumTx.AsTransaction().Hash(), height)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
 			},
 			true,
 			true,
@@ -1070,9 +1206,73 @@ func (suite *BackendTestSuite) TestGetEthBlockFromTendermint() {
 
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				RegisterConsensusParams(client, height)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetByTxHash(indexer, msgEthereumTx.AsTransaction().Hash(), height)
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, height)
 			},
 			true,
 			true,
+		},
+		{
+			name:      "pass - indexer returns error when getting latest block number",
+			baseFee:   sdk.NewInt(1).BigInt(),
+			validator: sdk.AccAddress(utiltx.GenerateAddress().Bytes()),
+			height:    int64(1),
+			resBlock: &tmrpctypes.ResultBlock{
+				Block: tmtypes.MakeBlock(1, []tmtypes.Tx{bz}, nil, nil),
+			},
+			blockRes: &tmrpctypes.ResultBlockResults{
+				Height:     1,
+				TxsResults: []*types.ResponseDeliverTx{{Code: 0, GasUsed: 0}},
+			},
+			fullTx: true,
+			registerMock: func(baseFee math.Int, validator sdk.AccAddress, height int64) {
+				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
+				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
+
+				client := suite.backend.clientCtx.Client.(*mocks.Client)
+				RegisterConsensusParams(client, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetByTxHash(indexer, msgEthereumTx.AsTransaction().Hash(), height)
+				RegisterIndexerGetLastRequestIndexedBlockErr(indexer)
+			},
+			expTxs:  true,
+			expPass: true,
+		},
+		{
+			name:      "fail - indexer returns error when getting tx by hash",
+			baseFee:   sdk.NewInt(1).BigInt(),
+			validator: sdk.AccAddress(utiltx.GenerateAddress().Bytes()),
+			height:    int64(1),
+			resBlock: &tmrpctypes.ResultBlock{
+				Block: tmtypes.MakeBlock(1, []tmtypes.Tx{bz}, nil, nil),
+			},
+			blockRes: &tmrpctypes.ResultBlockResults{
+				Height:     1,
+				TxsResults: []*types.ResponseDeliverTx{{Code: 0, GasUsed: 0}},
+			},
+			fullTx: false,
+			registerMock: func(baseFee math.Int, validator sdk.AccAddress, height int64) {
+				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
+				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
+
+				client := suite.backend.clientCtx.Client.(*mocks.Client)
+				RegisterConsensusParams(client, height)
+
+				RegisterParamsWithoutHeader(queryClient, height)
+
+				indexer := suite.backend.indexer.(*mocks.EVMTxIndexer)
+				RegisterIndexerGetByTxHashErr(indexer, msgEthereumTx.AsTransaction().Hash())
+				RegisterIndexerGetLastRequestIndexedBlock(indexer, 1)
+			},
+			expTxs:  false,
+			expPass: false,
 		},
 	}
 	for _, tc := range testCases {
@@ -1082,50 +1282,43 @@ func (suite *BackendTestSuite) TestGetEthBlockFromTendermint() {
 
 			block, err := suite.backend.RPCBlockFromTendermintBlock(tc.resBlock, tc.blockRes, tc.fullTx)
 
+			if tc.expPass {
+				suite.Require().NoError(err)
+			} else {
+				suite.Require().Error(err)
+				return
+			}
+
 			var expBlock map[string]interface{}
 			header := tc.resBlock.Block.Header
 			gasLimit := int64(^uint32(0)) // for `MaxGas = -1` (DefaultConsensusParams)
 			gasUsed := new(big.Int).SetUint64(uint64(tc.blockRes.TxsResults[0].GasUsed))
 
-			root := common.Hash{}.Bytes()
-			receipt := ethtypes.NewReceipt(root, false, gasUsed.Uint64())
-			bloom := ethtypes.CreateBloom(ethtypes.Receipts{receipt})
-
-			ethRPCTxs := []interface{}{}
+			var transactions ethtypes.Transactions
+			var receipts ethtypes.Receipts
 
 			if tc.expTxs {
-				if tc.fullTx {
-					rpcTx, err := ethrpc.NewRPCTransaction(
-						msgEthereumTx.AsTransaction(),
-						common.BytesToHash(header.Hash()),
-						uint64(header.Height),
-						uint64(0),
-						tc.baseFee,
-						suite.backend.chainID,
-					)
-					suite.Require().NoError(err)
-					ethRPCTxs = []interface{}{rpcTx}
-				} else {
-					ethRPCTxs = []interface{}{common.HexToHash(msgEthereumTx.Hash)}
-				}
+				transactions = append(transactions, msgEthereumTx.AsTransaction())
+				receipt := createTestReceipt(nil, tc.resBlock, msgEthereumTx, false, mockGasUsed)
+				receipts = append(receipts, receipt)
 			}
+
+			bloom := ethtypes.CreateBloom(receipts)
 
 			expBlock = ethrpc.FormatBlock(
 				header,
+				suite.backend.chainID,
 				tc.resBlock.Block.Size(),
-				gasLimit,
-				gasUsed,
-				ethRPCTxs,
+				gasLimit, gasUsed, tc.baseFee,
+				transactions, tc.fullTx,
+				receipts,
 				bloom,
 				common.BytesToAddress(tc.validator.Bytes()),
-				tc.baseFee,
+				log.NewNopLogger(),
 			)
 
 			if tc.expPass {
-				suite.Require().Equal(expBlock, block)
-				suite.Require().NoError(err)
-			} else {
-				suite.Require().Error(err)
+				suite.Equal(expBlock, block)
 			}
 		})
 	}