+++ rpc/types/utils.go
@@ -3,7 +3,11 @@ package types
 import (
 	"context"
 	"fmt"
+	"github.com/cometbft/cometbft/libs/log"
 	tmrpcclient "github.com/cometbft/cometbft/rpc/client"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/trie"
+	"github.com/pkg/errors"
 	"math/big"
 	"strings"
 
@@ -102,15 +106,55 @@ func BlockMaxGasFromConsensusParams(goCtx context.Context, clientCtx client.Cont
 // FormatBlock creates an ethereum block from a tendermint header and ethereum-formatted
 // transactions.
 func FormatBlock(
-	header tmtypes.Header, size int, gasLimit int64,
-	gasUsed *big.Int, transactions []interface{}, bloom ethtypes.Bloom,
-	validatorAddr common.Address, baseFee *big.Int,
+	header tmtypes.Header,
+	chainID *big.Int,
+	size int,
+	gasLimit int64, gasUsed *big.Int, baseFee *big.Int,
+	transactions ethtypes.Transactions, fullTx bool,
+	receipts ethtypes.Receipts,
+	bloom ethtypes.Bloom,
+	validatorAddr common.Address,
+	logger log.Logger,
 ) map[string]interface{} {
 	var transactionsRoot common.Hash
 	if len(transactions) == 0 {
 		transactionsRoot = ethtypes.EmptyRootHash
 	} else {
-		transactionsRoot = common.BytesToHash(header.DataHash)
+		transactionsRoot = ethtypes.DeriveSha(transactions, trie.NewStackTrie(nil))
+	}
+
+	var receiptsRoot common.Hash
+	if len(receipts) == 0 {
+		receiptsRoot = ethtypes.EmptyRootHash
+	} else {
+		receiptsRoot = ethtypes.DeriveSha(receipts, trie.NewStackTrie(nil))
+	}
+
+	var txsList []interface{}
+
+	for txIndex, tx := range transactions {
+		if !fullTx {
+			txsList = append(txsList, tx.Hash())
+			continue
+		}
+
+		height := uint64(header.Height) //#nosec G701 -- checked for int overflow already
+		index := uint64(txIndex)        //#nosec G701 -- checked for int overflow already
+
+		rpcTx, err := NewRPCTransaction(
+			tx,
+			common.BytesToHash(header.Hash()),
+			height,
+			index,
+			baseFee,
+			chainID,
+		)
+		if err != nil {
+			logger.Error("NewRPCTransaction failed", "hash", tx.Hash().Hex(), "error", err.Error())
+			continue
+		}
+
+		txsList = append(txsList, rpcTx)
 	}
 
 	result := map[string]interface{}{
@@ -130,10 +174,10 @@ func FormatBlock(
 		"gasUsed":          (*hexutil.Big)(gasUsed),
 		"timestamp":        hexutil.Uint64(header.Time.Unix()),
 		"transactionsRoot": transactionsRoot,
-		"receiptsRoot":     ethtypes.EmptyRootHash,
+		"receiptsRoot":     receiptsRoot,
 
 		"uncles":          []common.Hash{},
-		"transactions":    transactions,
+		"transactions":    txsList,
 		"totalDifficulty": (*hexutil.Big)(big.NewInt(0)),
 	}
 
@@ -216,6 +260,72 @@ func NewRPCTransaction(
 		}
 	}
 	return result, nil
+}
+
+func NewRPCReceipt(
+	ethMsg *evmtypes.MsgEthereumTx,
+	transactionIndex hexutil.Uint64,
+	success bool,
+	gasUsed hexutil.Uint64,
+	cumulativeGasUsed hexutil.Uint64,
+	baseFee *big.Int,
+	logs []*ethtypes.Log,
+	blockHash common.Hash,
+	blockNumber hexutil.Uint64,
+	chainID *big.Int,
+) (receipt *RPCReceipt, err error) {
+	var status hexutil.Uint
+
+	if success {
+		status = hexutil.Uint(ethtypes.ReceiptStatusSuccessful)
+	} else {
+		status = hexutil.Uint(ethtypes.ReceiptStatusFailed)
+	}
+
+	if logs == nil {
+		logs = []*ethtypes.Log{}
+	}
+
+	from, err := ethMsg.GetSender(chainID)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to get sender")
+	}
+
+	txData, err := evmtypes.UnpackTxData(ethMsg.Data)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to unpack tx data")
+	}
+
+	rpcReceipt := RPCReceipt{
+		Status:            status,
+		CumulativeGasUsed: cumulativeGasUsed,
+		Bloom:             ethtypes.BytesToBloom(ethtypes.LogsBloom(logs)),
+		Logs:              logs,
+		TransactionHash:   ethMsg.AsTransaction().Hash(),
+		ContractAddress:   nil,
+		GasUsed:           gasUsed,
+		BlockHash:         blockHash,
+		BlockNumber:       blockNumber,
+		TransactionIndex:  transactionIndex,
+		Type:              hexutil.Uint(ethMsg.AsTransaction().Type()),
+		From:              from,
+		To:                txData.GetTo(),
+		EffectiveGasPrice: nil,
+	}
+
+	if rpcReceipt.To == nil {
+		newContractAddress := crypto.CreateAddress(from, txData.GetNonce())
+		rpcReceipt.ContractAddress = &newContractAddress
+	}
+
+	if baseFee != nil {
+		if dynamicTx, ok := txData.(*evmtypes.DynamicFeeTx); ok {
+			effectiveGasPrice := hexutil.Big(*dynamicTx.EffectiveGasPrice(baseFee))
+			rpcReceipt.EffectiveGasPrice = &effectiveGasPrice
+		}
+	}
+
+	return &rpcReceipt, nil
 }
 
 // BaseFeeFromEvents parses the feemarket basefee from cosmos events